requirements.txt
Adafruit-ADS1x15==1.0.2
Adafruit-GPIO==1.0.3
Adafruit-PureIO==0.2.1
asn1crypto==0.24.0
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
automationhat==0.1.0
bcrypt==3.1.4
beautifulsoup4==4.6.0
blinker==1.3
blinkt==0.1.2
Cap1xxx==0.1.3
cffi==1.11.4
chardet==2.3.0
click==6.6
colorama==0.3.7
constantly==15.1.0
cryptography==2.1.4
drumhat==0.1.0
envirophat==1.0.0
ExplorerHAT==0.4.2
Flask==0.12.1
fourletterphat==0.1.0
google==2.0.1
google-api-python-client==1.6.5
gpiozero==1.4.1
httplib2==0.10.3
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.8
keyring==10.1
keyrings.alt==1.3
MarkupSafe==0.23
mcpi==0.1.1
microdotphat==0.2.1
mock==2.0.0
mote==0.0.3
motephat==0.0.2
numpy==1.12.1
oauth2client==4.1.2
oauthlib==2.0.1
paramiko==2.4.0
pbr==3.1.1
pgzero==1.2
phatbeat==0.1.1
pi-ina219==1.1.0
pianohat==0.1.0
picamera==1.13
picraft==1.0
piglow==1.2.4
pigpio==1.38
Pillow==4.0.0
plumbum==1.6.5
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pycrypto==2.6.1
pygame==1.9.3
pygobject==3.22.0
pyinotify==0.9.6
PyJWT==1.4.2
PyNaCl==1.2.1
pyOpenSSL==16.2.0
pyserial==3.2.1
pysftp==0.2.9
python-apt==1.1.0b5
pyxdg==0.25
rainbowhat==0.1.0
requests==2.12.4
requests-oauthlib==0.7.0
RPi.GPIO==0.6.3
rpyc==3.4.4
rsa==3.4.2
RTIMULib==7.2.1
scrollphat==0.0.7
scrollphathd==1.2.1
SecretStorage==2.3.1
sense-emu==1.1
sense-hat==2.2.0
simplejson==3.10.0
six==1.11.0
skywriter==0.0.7
sn3218==1.2.7
spidev==3.2
thonny==2.1.16
tornado==4.5.3
touchphat==0.0.1
Twisted==17.9.0
twython==3.4.0
txaio==2.8.2
uritemplate==3.0.0
urllib3==1.19.1
Werkzeug==0.11.15
zope.interface==4.4.3
